+++ kubejs/client_scripts/JEI.js
@@ -5,13 +5,6 @@ JEIEvents.hideItems(event => {
 	event.hide(/sophisticatedstorage:(.*)shulker_box/)
 })
 
-JEIEvents.hideFluids(event => {
-	event.hide('ad_astra:oil')
-	event.hide('ad_astra:oxygen')
-	event.hide('productivebees:honey')
-	event.hide('productivebees:flowing_honey')
-})
-
 JEIEvents.addItems(event => {
 	event.add('minecraft:bucket')
 	event.add('minecraft:water_bucket')