+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 - Remove duplicate netherite material
 - Fix tagprefix ignores for stone material
 - Hide sophisticated backpacks/storage crafting upgrade workstations in EMI
-
+- Fix fluids not hiding in EMI
 **Mod Changes**
 -